requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.6
django-crispy-forms==2.1
django-extensions==3.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
